feat(jwt): Cleanup
