--- module/auth-model/index.ts
@@ -1,2 +1,0 @@
-export * from './src/principal';
-export * from './src/service';
