--- module/auth-model/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@travetto/auth-model",
-  "displayName": "Model Auth Source",
-  "version": "2.0.0-alpha.3",
-  "description": "Model-based authentication and registration support for the travetto framework",
-  "keywords": [
-    "authentication",
-    "model",
-    "travetto",
-    "registration",
-    "typescript"
-  ],
-  "homepage": "https://travetto.io",
-  "license": "MIT",
-  "author": {
-    "email": "travetto.framework@gmail.com",
-    "name": "Travetto Framework"
-  },
-  "files": [
-    "index.ts",
-    "src",
-    "test-support"
-  ],
-  "main": "index.ts",
-  "repository": {
-    "url": "https://github.com/travetto/travetto.git",
-    "directory": "module/auth-model"
-  },
-  "dependencies": {
-    "@travetto/auth": "^2.0.0-alpha.2",
-    "@travetto/model": "^2.0.0-alpha.3"
-  },
-  "optionalPeerDependencies": {
-    "@travetto/auth-rest": "^2.0.0-alpha.4"
-  },
-  "docDependencies": {
-    "@travetto/app": "^2.0.0-alpha.3"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
