+++ module/auth/doc/model/config.ts
@@ -1,5 +1,5 @@
 import { InjectableFactory } from '@travetto/di';
-import { ModelAuthService, RegisteredPrincipal } from '@travetto/auth-model';
+import { ModelAuthService, RegisteredPrincipal } from '@travetto/auth';
 import { ModelCrudSupport } from '@travetto/model';
 
 import { User } from './model';
