+++ module/auth/index.ts
@@ -1,2 +1,5 @@
 export * from './src/util';
 export * from './src/types';
+
+// Named export needed for proxying
+export * from './src/extension/model';
