+++ module/auth/package.json
@@ -16,7 +16,8 @@
   },
   "files": [
     "index.ts",
-    "src"
+    "src",
+    "test-support"
   ],
   "main": "index.ts",
   "repository": {
@@ -25,6 +26,9 @@
   },
   "dependencies": {
     "@travetto/base": "^2.0.0-alpha.2"
+  },
+  "optionalPeerDependencies": {
+    "@travetto/model": "^2.0.0-alpha.3"
   },
   "publishConfig": {
     "access": "public"
