+++ module/auth/src/internal/types.ts
@@ -1,3 +1,7 @@
+import { ContextProvider } from '@travetto/rest';
+
+@ContextProvider((c, req) => req.auth)
 export class PrincipalTarget { }
+
 export class AuthorizerTarget { }
 export class AuthenticatorTarget { }
