+++ module/jwt/index.ts
@@ -1,7 +1,6 @@
-export * from './src/decode';
 export * from './src/verify';
 export * from './src/util';
 export * from './src/types';
 export * from './src/error';
 // Named export needed for proxying
-export { JWTAuthContextEncoder } from './src/extension/auth-rest';
+export { JWTPrincipalEncoder } from './src/extension/auth-rest';
