--- related/travetto.github.io/src/app/documentation/gen/auth-model/auth-model.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-auth-model',
-  templateUrl: './auth-model.component.html',
-  styleUrls: []
-})
-export class AuthModelComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
